web_server/requirements/base.txt
fastapi>=0.75.0, <1.0.0
uvicorn>=0.17.6, <0.18.0
pyyaml>=6.0, <7.0.0
requests>=2.27.1, <2.28.0
paho-mqtt>=1.5.1